prod page carousels temp title fix / version
Chrome was removing the nth-child(3) pseudo selector from the title
carousel on click of the next button. This caused the element to land
under the button instead of to the right of it.
Possible that a call back function in the caroufredsel plugin will take
care of reasigning the properties for the middle element so the design
can be matched.